requirements.txt
# Web 框架
fastapi>=0.103.0,<0.104.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
email-validator>=1.1.3
itsdangerous>=2.1.2

# 数据库
pymongo>=4.0.1
motor>=2.5.1  # 异步 MongoDB 驱动

# 安全
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4,<1.8.0  # 密码哈希
bcrypt>=3.2.0,<4.0.0  # 确保与passlib兼容的bcrypt版本
python-multipart>=0.0.5  # 表单处理

# 环境变量
python-dotenv>=0.19.0

# Spotify API
spotipy>=2.19.0
spotdl>=4.2.9  # 指定兼容的 spotdl 版本

# 工具
tenacity>=8.0.1  # 重试机制
requests>=2.26.0
aiofiles>=0.7.0  # 异步文件操作
python-slugify>=5.0.2  # URL 友好的字符串

# 日志
loguru>=0.5.3